Toy with setup.sh

+++ setup.sh
@@ -1,16 +1,24 @@
 #!/usr/bin/env bash
 
 set -e # Abort on error
-set -u # Error on unset varibles
 set -o pipefail
 
 # install Plug package manager
-curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [ ! -f ~/.vim/autoload/plug.vim ]
+then
+  curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+fi
 
+if [ -z "${XDG_CONFIG_HOME}" ]; then
+  XDG_CONFIG_HOME=$HOME/.config
+fi
 # Symlink neovim config into place
-mkdir -p ${XDG_CONFIG_HOME:=$HOME/.config}
-ln -s ~/.vim $XDG_CONFIG_HOME/nvim
-ln -s ~/.vimrc $XDG_CONFIG_HOME/nvim/init.vim
+if [ ! -f $XDG_CONFIG_HOME/nvim/init.vim ]
+then
+  mkdir -p $XDG_CONFIG_HOME
+  ln -s ~/.vim $XDG_CONFIG_HOME/nvim
+  ln -s ~/.vimrc $XDG_CONFIG_HOME/nvim/init.vim
+fi
 
 # Setup undo history
 mkdir -p ~/.vim/undo/
@@ -21,3 +29,5 @@ if [ "$(uname)" == "Darwin" ]; then
   tic $TERM.ti
   cd -
 fi
+
+echo Now run 'pip install neovim'
